Allow trailing comma in class static vars
